+++ apps/meteor/client/views/admin/sidebarItems.ts
@@ -14,6 +14,12 @@ export const {
 		permissionGranted: (): boolean => hasPermission('view-statistics'),
 	},
 	{
+		href: '/admin/subscription',
+		i18nLabel: 'Subscription',
+		icon: 'card',
+		permissionGranted: (): boolean => hasPermission('manage-cloud'),
+	},
+	{
 		href: '/admin/registration',
 		i18nLabel: 'Registration',
 		icon: 'cloud-plus',