+++ packages/rest-typings/src/v1/omnichannel.ts
@@ -765,7 +765,7 @@ const LivechatUsersManagerGETSchema = {
 			nullable: true,
 		},
 		onlyAvailable: {
-			type: 'string',
+			type: 'boolean',
 			nullable: true,
 		},
 		excludeId: {
@@ -3277,7 +3277,7 @@ export type OmnichannelEndpoints = {
 		}>;
 	};
 	'/v1/livechat/tags/:tagId': {
-		GET: () => ILivechatTag | null;
+		GET: () => ILivechatTag;
 	};
 	'/v1/livechat/department': {
 		GET: (params?: LivechatDepartmentProps) => PaginatedResult<{
@@ -3615,7 +3615,7 @@ export type OmnichannelEndpoints = {
 		POST: (params: POSTLivechatTriggersParams) => void;
 	};
 	'/v1/livechat/triggers/:_id': {
-		GET: () => { trigger: ILivechatTrigger | null };
+		GET: () => { trigger: ILivechatTrigger };
 		DELETE: () => void;
 	};
 	'/v1/livechat/rooms': {