+++ packages/ui-contexts/CHANGELOG.md
@@ -1,5 +1,45 @@
 # @rocket.chat/ui-contexts
 
+## 2.0.5
+
+### Patch Changes
+
+- @rocket.chat/core-typings@6.4.5
+- @rocket.chat/rest-typings@6.4.5
+- @rocket.chat/ddp-client@0.2.5
+
+## 2.0.4
+
+### Patch Changes
+
+- @rocket.chat/core-typings@6.4.4
+- @rocket.chat/rest-typings@6.4.4
+- @rocket.chat/ddp-client@0.2.4
+
+## 2.0.3
+
+### Patch Changes
+
+- @rocket.chat/core-typings@6.4.3
+- @rocket.chat/rest-typings@6.4.3
+- @rocket.chat/ddp-client@0.2.3
+
+## 2.0.2
+
+### Patch Changes
+
+- @rocket.chat/core-typings@6.4.2
+- @rocket.chat/rest-typings@6.4.2
+- @rocket.chat/ddp-client@0.2.2
+
+## 2.0.1
+
+### Patch Changes
+
+- @rocket.chat/core-typings@6.4.1
+- @rocket.chat/rest-typings@6.4.1
+- @rocket.chat/ddp-client@0.2.1
+
 ## 2.0.0
 
 ### Minor Changes